+++ app/helpers/agilibox/bootstrap_helper.rb
@@ -13,8 +13,8 @@ module Agilibox::BootstrapHelper
   end
 
   def bs_progress_bar(percentage)
-    content_tag(:div, class: "progress") do
-      content_tag(:div, class: "progress-bar", style: "width:#{percentage}%") do
+    tag.div(class: "progress") do
+      tag.div(class: "progress-bar", style: "width:#{percentage}%") do
         "#{percentage}%"
       end
     end