+++ app/helpers/agilibox/filters_helper.rb
@@ -7,7 +7,7 @@ module Agilibox::FiltersHelper
     text = options.delete(:text) || t("actions.filter")
     icon = options.delete(:icon) || :filter
 
-    content_tag(:button, options) do
+    tag.button(options) do
       icon(icon) + " " + text
     end
   end
@@ -20,7 +20,7 @@ module Agilibox::FiltersHelper
     text = options.delete(:text) || t("actions.reset")
     icon = options.delete(:icon) || :undo
 
-    content_tag(:button, options) do
+    tag.button(options) do
       icon(icon) + " " + text
     end
   end