+++ lib/agilibox/active_model_type_cast.rb
@@ -2,7 +2,7 @@ module Agilibox::ActiveModelTypeCast
   module Decimal
     def cast_value(value)
       if value.is_a?(String)
-        super value.tr(",", ".").gsub(/[^-0-9\.]/, "")
+        super value.tr(",", ".").gsub(/[^-0-9.]/, "")
       else
         super value
       end