+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## Next version
+- Fix info helper with false
+- Update Rubocop + fix offences
 
 ## 1.9.5
 - Fix warnings on Rails 6